Search .env file in parent directories
Extracts the methods responsible for this into their own module and uses this
module for searching yml and .env files. This follows the same rules as the yml
files.

So you can have base_dir/.env and no project_dir/.env or project_dir/.env.
Before .env file was only used if you invoke from the base directory.
